==> notebooks/00_split_data.ipynb <==
import pandas as pd

df = pd.read_csv("../data/raw/house_prices_raw.csv")

pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)


df.shape

df.head(2)

# Range of years
df['date'] = pd.to_datetime(df['date'])
print(df['date'].dt.year.value_counts().sort_index())

df = df.sort_values("date")

# cutoff date for train/test split
cutoff_date_eval = "2020-01-01"
cutoff_date_holdout = "2022-01-01"

# Train: all data before cutoff_date_eval
df_train = df[df['date'] < cutoff_date_eval]

# Eval: all data between cutoff_date_eval and cutoff_date_holdout
df_eval = df[(df['date'] >= cutoff_date_eval) & (df['date'] < cutoff_date_holdout)]

# Holdout: all data after cutoff_date_holdout
df_holdout = df[df['date'] >= cutoff_date_holdout]

print("Train set: ", df_train.shape)
print("Eval set:", df_eval.shape)
print("Holdout set:", df_holdout.shape)

df_train.head(2)


df_eval.head(2)


df_holdout.head(2)

#SAVE
df_train.to_csv("../data/raw/house_prices_train.csv", index=False)
df_eval.to_csv("../data/raw/house_prices_eval.csv", index=False)
df_holdout.to_csv("../data/raw/house_prices_holdout.csv", index=False)
