==> notebooks/02_engineering_features.ipynb <==
import pandas as pd
from category_encoders import TargetEncoder

train_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/train_cleaned.csv")
eval_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/eval_cleaned.csv")
holdout_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/holdout_cleaned.csv")

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

print(train_df["date"].min(), "to", train_df["date"].max())
print(eval_df["date"].min(), "to", eval_df["date"].max())

train_df["date"] = pd.to_datetime(train_df["date"])
eval_df["date"] = pd.to_datetime(eval_df["date"])
holdout_df["date"] = pd.to_datetime(holdout_df["date"])

train_df.head(5)


def date_feature(df: pd.DataFrame) -> pd.DataFrame:
    """
    Splits the date column into individual year, month, and day columns
    and inserts them into the dataframe in the order of year, month, day.
    """
    df["year"] = df["date"].dt.year
    df["month"] = df["date"].dt.month
    df["day"] = df["date"].dt.day

    df.insert(1, "year", df.pop("year"))
    df.insert(2, "month", df.pop("month"))
    df.insert(3, "day", df.pop("day"))
    return df

train_df = date_feature(train_df)
eval_df = date_feature(eval_df)
holdout_df = date_feature(holdout_df)

train_df.head(1)

eval_df.head(2)

"""
Frequency Encoding:
 - For Zipcodes
 - For each zipcode, calculate the frequency of that zipcode in the training data 
   and replace the zipcode with that frequency.
"""

# Create frequency dict
zipcode_freq = train_df["zipcode"].value_counts() 

train_df["zipcode_freq"] = train_df["zipcode"].map(zipcode_freq)

# Don't compute frequency on evaluation
eval_df["zipcode_freq"] = eval_df["zipcode"].map(zipcode_freq).fillna(0)

holdout_df["zipcode_freq"] = holdout_df["zipcode"].map(zipcode_freq).fillna(0)

print(train_df[["zipcode", "zipcode_freq"]].head())



train_df.head(1)

"""
Target Encoding:
For each city, encode the city with the avg price 
"""
te = TargetEncoder(cols=["city_full"])
train_df["encoded_city"] = te.fit_transform(train_df["city_full"], train_df["price"])
eval_df["encoded_city"] = te.transform(eval_df["city_full"])
holdout_df["encoded_city"] = te.transform(holdout_df["city_full"])

print(train_df[["city_full", "encoded_city"]].head())


# REMOVE UNUSED COLUMNS TO PREVENT DATA LEAKAGE
drop = ["date", "city_full", "zipcode", "city", "median_sale_price"]
train_df.drop(columns=drop, inplace=True)
eval_df.drop(columns=drop, inplace=True)
holdout_df.drop(columns=drop, inplace=True)

print(train_df.shape)
train_df.head(1)

train_df.to_csv("/Users/larry/house-price-regression/data/processed/train_engineered.csv", index=False)
eval_df.to_csv("/Users/larry/house-price-regression/data/processed/eval_engineered.csv", index=False)
holdout_df.to_csv("/Users/larry/house-price-regression/data/processed/holdout_engineered.csv", index=False)
