==> notebooks/03_baseline.ipynb <==
import pandas as pd
from sklearn.dummy import DummyRegressor
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np

train_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/train_engineered.csv")
eval_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/eval_engineered.csv")

target = "price"
X_train = train_df.drop(columns=target)
y_train = train_df[target]

X_eval = eval_df.drop(columns=target)
y_eval = eval_df[target]

baseline_model = DummyRegressor(strategy = "median")
baseline_model.fit(X_train, y_train)

y_prediction = baseline_model.predict(X_eval)

mae = mean_absolute_error(y_eval, y_prediction)
rmse = np.sqrt(mean_squared_error(y_eval, y_prediction))
r2 = r2_score(y_eval, y_prediction)

print(f"Baseline MAE: {mae}")
print(f"Baseline RMSE: {rmse}")
print(f"Baseline R2: {r2}")
