==> notebooks/04_linear_regression.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np

train_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/train_engineered.csv")
eval_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/eval_engineered.csv")


target = "price"
X_train = train_df.drop(columns=target)
y_train = train_df[target]

X_eval = eval_df.drop(columns=target)
y_eval = eval_df[target]

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_eval_scaled = scaler.transform(X_eval)


# TRAINING
linear_model = LinearRegression()
linear_model.fit(X_train_scaled, y_train)
# EVALUATION
y_pred_linear = linear_model.predict(X_eval_scaled)

print("Linear Regression Performance:")
print(f"MAE: {mean_absolute_error(y_eval, y_pred_linear)}")
print(f"RMSE: {np.sqrt(mean_squared_error(y_eval, y_pred_linear))}")
print(f"R2: {r2_score(y_eval, y_pred_linear)}")
