==> notebooks/05_xgboost.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from xgboost import XGBRegressor



train_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/train_engineered.csv")
eval_df = pd.read_csv("/Users/larry/house-price-regression/data/processed/eval_engineered.csv")

target = "price"

X_train = train_df.drop(columns=target)
y_train = train_df[target]  

X_eval = eval_df.drop(columns=target)
y_eval = eval_df[target]

xgboost_model = XGBRegressor(
    n_estimators=500,
    learning_rate=0.01,
    max_depth=6,
    subsample=0.8,
    colsample_bytree=0.8,
    random_state=42,
    n_jobs=-1
)

xgboost_model.fit(X_train, y_train)

y_pred_xgboost = xgboost_model.predict(X_eval)

mae = mean_absolute_error(y_eval, y_pred_xgboost)
rmse = np.sqrt(mean_squared_error(y_eval, y_pred_xgboost))
r2 = r2_score(y_eval, y_pred_xgboost)

print("XGBoost Performance:")
print(f"MAE: {mae}")
print(f"RMSE: {rmse}")
print(f"R2: {r2}")
